requirements.txt
ansicolors>=1.0.2,<2.0
Django>=1.9,<2.0
newrelic>=2.60.0.46,<3.0
psycopg2>=2.6.1,<3.0
requests>=2.8.1,<3.0
